Feature/init request and comment

+++ src/backend/SomeSandwich.FakeMentorus.Domain/Grade/Grade.cs
@@ -5,53 +5,53 @@ using SomeSandwich.FakeMentorus.Domain.Users;
 namespace SomeSandwich.FakeMentorus.Domain.Grade;
 
 /// <summary>
-///     Grade entity.
+/// Grade entity.
 /// </summary>
 public class Grade
 {
     /// <summary>
-    ///     Grade id.
+    /// Grade id.
     /// </summary>
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public int Id { get; set; }
 
     /// <summary>
-    ///     Grade Composition id.
+    /// Grade Composition id.
     /// </summary>
     required public int GradeCompositionId { get; set; }
 
     /// <summary>
-    ///     Student id.
+    /// Student id.
     /// </summary>
     required public int StudentId { get; set; }
 
     /// <summary>
-    ///     Grade value.
+    /// Grade value.
     /// </summary>
-    required public int GradeValue { get; set; }
+    required public float GradeValue { get; set; }
 
     // ---------------------------------------------------------------------------------------------
 
     /// <summary>
-    ///     Grade comment.
+    /// Grade comment.
     /// </summary>
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
 
     /// <summary>
-    ///     Updated at.
+    /// Updated at.
     /// </summary>
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 
     // ---------------------------------------------------------------------------------------------
 
     /// <summary>
-    ///     Grade Composition.
+    /// Grade Composition.
     /// </summary>
     public virtual GradeComposition GradeComposition { get; set; } = null!;
 
     /// <summary>
-    ///     Student.
+    /// Student.
     /// </summary>
     public virtual User Student { get; set; } = null!;
 }
